fix(circular_buffer): Fixed the return of empty string
The circular buffer returned an empty string allocated with a size of BUFFER_SIZE - 1 if the last command contained a "\n". the buffer->tail was correctly set but on the search of the substring the buffer located the old "\n".
To fix it quickly i just removed the "\n" in the buffer